requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
transformers
torch
Pillow
faiss-cpu
